+++ spec/chewy/search/parameters/inner_hits_spec.rb
@@ -1,0 +1,5 @@
+require 'chewy/search/parameters/hash_storage_examples'
+
+describe Chewy::Search::Parameters::InnerHits do
+  it_behaves_like :hash_storage, :inner_hits
+end